add favBtn hover stars on product page
